Move monotonicity to function_monotonic
