+++ common/functions/src/scalars/mod.rs
@@ -22,6 +22,7 @@ mod function_alias;
 mod function_column;
 mod function_factory;
 mod function_literal;
+mod function_monotonic;
 mod hashes;
 mod logics;
 mod maths;
@@ -37,11 +38,11 @@ pub use conditionals::*;
 pub use dates::*;
 pub use expressions::*;
 pub use function::Function;
-pub use function::Monotonicity;
 pub use function_alias::AliasFunction;
 pub use function_column::ColumnFunction;
 pub use function_factory::FunctionFactory;
 pub use function_literal::LiteralFunction;
+pub use function_monotonic::Monotonicity;
 pub use hashes::*;
 pub use logics::*;
 pub use maths::*;
